adjusted order tests

+++ src/models/orders.ts
@@ -3,10 +3,15 @@ import client from '../database';
 
 export type Order = {
   id?: number;
-  product_id: number;
-  quantity: number;
   user_id: number;
+  quantity: number;
   status: string;
+};
+
+export type OrderProduct = {
+  id?: number;
+  order_id?: number;
+  product_id: number;
 };
 
 export class OrderStore {
@@ -15,7 +20,7 @@ export class OrderStore {
     const conn = await client.connect();
 
     try {
-      const sql = 'SELECT * FROM orders';
+      const sql = 'SELECT o.*, op.product_id, op.quantity FROM Orders o JOIN order_product op ON o.id = op.order_id';
       const result = await conn.query(sql);
 
       return result.rows;
@@ -31,7 +36,8 @@ export class OrderStore {
     const conn = await client.connect();
 
     try {
-      const sql = 'SELECT * FROM orders WHERE id=($1)';
+      const sql =
+        'SELECT o.id, o.quantity, o.status FROM Orders o JOIN order_product op ON o.id = op.order_id WHERE o.user_id = ($1)';
       const result = await conn.query(sql, [id]);
 
       return result.rows[0];
@@ -42,12 +48,6 @@ export class OrderStore {
     }
   }
 
-  //   id SERIAL PRIMARY KEY,
-  // order_id INT,
-  // product_id INT,
-  // quantity INTEGER,
-  // FOREIGN KEY (order_id) REFERENCES Orders(id),
-  // FOREIGN KEY (product_id) REFERENCES Products(id)
   async userOrders(id: string, status: string | null): Promise<Order[]> {
     // @ts-ignore
     const conn = await client.connect();
@@ -57,10 +57,12 @@ export class OrderStore {
       let params: string[];
 
       if (status) {
-        sql = 'SELECT * FROM order_product WHERE user_id=($1) AND status=($2)';
+        sql =
+          'SELECT o.id, o.quantity, o.status FROM Orders o JOIN order_product op ON o.id = op.order_id WHERE o.user_id = ($1) AND o.status = ($2)';
         params = [id, status];
       } else {
-        sql = 'SELECT * FROM orders WHERE user_id=($1)';
+        sql =
+          'SELECT o.id, o.quantity, o.status FROM Orders o JOIN order_product op ON o.id = op.order_id WHERE o.user_id = ($1)';
         params = [id];
       }
 
@@ -74,16 +76,24 @@ export class OrderStore {
     }
   }
 
-  async create(b: Order): Promise<Order> {
+  async create(b: Order & OrderProduct): Promise<OrderProduct | Order> {
     // @ts-ignore
     const conn = await client.connect();
 
     try {
-      const sql = 'INSERT INTO orders (product_id, quantity, user_id, status) VALUES($1, $2, $3, $4) RETURNING *';
-
-      const result = await conn.query(sql, [b.product_id, b.quantity, b.user_id, b.status]);
-
+      // First create an order in 'orders' table
+      const sql = 'INSERT INTO orders (user_id, quantity, status) VALUES($1, $2, $3) RETURNING *';
+      const result = await conn.query(sql, [b.user_id, b.quantity, b.status]);
       const order = result.rows[0];
+
+      if (order) {
+        // Then create add product details with order.id in 'order_product's
+        const sql = 'INSERT INTO order_product (order_id, product_id) VALUES($1, $2) RETURNING *';
+        const result = await conn.query(sql, [order.id, b.product_id]);
+        const productOrder = result.rows[0];
+
+        return productOrder;
+      }
 
       return order;
     } catch (err) {
@@ -93,16 +103,25 @@ export class OrderStore {
     }
   }
 
-  async delete(id: string): Promise<Order> {
+  async delete(id: string): Promise<Order | OrderProduct> {
     // @ts-ignore
     const conn = await client.connect();
 
     try {
-      const sql = 'DELETE FROM orders WHERE id=($1)';
+      // first delete product_order based on order_id
+      const sqlProductOrders = 'DELETE FROM order_product WHERE order_id=($1)';
+      const resultProductOrders = await conn.query(sqlProductOrders, [id]);
 
-      const result = await conn.query(sql, [id]);
+      if (resultProductOrders) {
+        // then do the filing
+        const sqlOrders = 'DELETE FROM orders WHERE id=($1)';
 
-      return result;
+        const resultOrders = await conn.query(sqlOrders, [id]);
+
+        return resultOrders;
+      }
+
+      return resultProductOrders;
     } catch (err) {
       throw new Error(`Could not delete order ${id}. Error: ${err}`);
     } finally {
